requirements.txt
selenium==4.12.0
Pillow==10.1.0
python-telegram-bot==20.5
urllib3==2.0.5
certifi==2025.7.1
